refactor: lograr abstracción eficiente de BaseAuditModel
- se extiende desde core a otras apps
