+++ core/models.py
@@ -1,3 +1,10 @@
 from django.db import models
 
-# Create your models here.
+class BaseAuditModel(models.Model):
+    is_active = models.BooleanField(default=True)  # Soft delete
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    deleted_at = models.DateTimeField(null=True, blank=True)
+
+    class Meta:
+        abstract = True
