+++ finance/models.py
@@ -1,14 +1,6 @@
 from django.db import models
 from rooms.models import TipoHabitacion
-
-class BaseAuditModel(models.Model):
-    is_active = models.BooleanField(default=True)  # Soft delete
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    deleted_at = models.DateTimeField(null=True, blank=True)
-
-    class Meta:
-        abstract = True
+from core.models import BaseAuditModel
 
 class Temporada(BaseAuditModel):
     class Nombre(models.TextChoices):
