+++ restaurant/models.py
@@ -1,14 +1,5 @@
 from django.db import models
-
-class BaseAuditModel(models.Model):
-    is_active = models.BooleanField(default=True)  # Soft delete
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    deleted_at = models.DateTimeField(null=True, blank=True)
-
-    class Meta:
-        abstract = True
-
+from core.models import BaseAuditModel
 # === Mesa ===
 class Mesa(BaseAuditModel):
     numero_mesa = models.SmallIntegerField()
